feat: train image models in Scratch automatically
Models can't be shared between the training site and Scratch for
image projects any more (because of the move to subdomains). This
is a change in behaviour that is causing confusion. To mitigate
this, this commit will get the Scratch extension to try and train
it's own model on startup.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>