fix: add workaround for Python apps due to Cloudflare issue
Cloudflare is blocking access to the Machine Learning for Kids API
for user agents that it doesn't think look like a web browser.

I am chasing this with them to get them to stop doing this, but in
the meantime this commit provides students with a workaround, by
getting them to bypass Cloudflare and access one of the site
instances directly.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>