chore: url changes to avoid caching issues
There are a small number of users who are still reporting CORS
issues, because they're not able to clear their browser cache of
the previously cached resources without CORS headers.

Rather than keep trying to explain how to clear a browser cache,
it seems like the simplest fix for these users is just to move the
resources to a new url.
To avoid inventing a whole new set of caching problems, I'm
leaving the existing resources where they are as well.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>