fix: improvements to multi-region / failover / geo-steering
The main fix is in deciding when to failover to a different region
as the previous check was failing on any non-2xx response code. So
correctly-handled-but-bad requests (e.g. 400s, 401s, 403s, 404s)
were all getting sprayed to all three regions.

I'm also trying to address some annoying errors facing Python and
App Inventor users, by trying to inject user-agents for them when
they don't include them for themselves. I'm not sure that's
working though - it's more an experiment.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>