fix: add workaround for AppInventor due to Cloudflare issue
Same problem as 5e64f09369869e1e3738306c9508a64a1470aa4d

For some reason, Cloudflare is blocking what it thinks are
suspicious requests to the Machine Learning for Kids API (where
"suspicious" means "doesn't look like a web browser") and this
is blocking App Inventor users. The result is they get an
http-403 permission denied error.

I'm still chasing on how to get this unblocked, but in the
meantime this commit provides a custom build of the App Inventor
extension that uses the same temporary alternative site URL that
I'm providing for Python users in
5e64f09369869e1e3738306c9508a64a1470aa4d

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>