feat: use some data from datasets for testing
This commit adds a new feature: a new option when importing a
project from an existing dataset. This lets the user exclude some
of the dataset from the training data, and make it available for
use in testing.

Test data is returned in the POST import API response, and saved
in the browser using localStorage. From there, the user can
download it to a CSV file from the "Learn & Test" page.

Note: This does introduce a limitation - users without
localStorage support in their browser (most notably, Safari when
used in Private Browsing mode) won't be able to save their test
data. This felt like an acceptable compromise to avoid storing a
lot of additional data in the server.

The commit also introduces three new datasets (which introduced
the requirement for this test data feature).

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>