chore: reduce size of fake news dataset
I'm nervous about the storage costs that could result if a large
number of users import the Fake News dataset as it's pretty big.

This commit reduces it to 1000 examples (500 real, 500 fake),
which is still big but at least a reduction from the 1465 examples
in there before.

I'll keep an eye on the impact this project has, and reduce the
dataset further if it looks problematic.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>