feat: use some data from datasets for testing (#469)
* feat: use some data from datasets for testing

This commit adds a new feature: a new option when importing a
project from an existing dataset. This lets the user exclude some
of the dataset from the training data, and make it available for
use in testing.

Test data is returned in the POST import API response, and saved
in the browser using localStorage. From there, the user can
download it to a CSV file from the "Learn & Test" page.

Note: This does introduce a limitation - users without
localStorage support in their browser (most notably, Safari when
used in Private Browsing mode) won't be able to save their test
data. This felt like an acceptable compromise to avoid storing a
lot of additional data in the server.

The commit also introduces three new datasets (which introduced
the requirement for this test data feature).

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>

* chore: reduce size of fake news dataset

I'm nervous about the storage costs that could result if a large
number of users import the Fake News dataset as it's pretty big.

This commit reduces it to 1000 examples (500 real, 500 fake),
which is still big but at least a reduction from the 1465 examples
in there before.

I'll keep an eye on the impact this project has, and reduce the
dataset further if it looks problematic.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>

* fix: fix travis build

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>