chore: removing dead code
Creating a zip file of images was required when the site used the
IBM Watson Visual Recognition service - the API for training a new
model needed a zip file to be uploaded.

The Visual Recognition service has long since been retired, so this
code hasn't been used in a long time. Rather than continue to
maintain it, it's easiest to delete it.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>