fix: migrate from request to got for image downloads
A large increase in students doing image projects has led to an
increase in ETIMEDOUT and ESOCKETTIMEDOUT errors when downloading
images for verifying and resizing.

It appears that this is due to blocking DNS lookups, as a result
of the fact that the image downloading is naturally performing
http requests to a large number of different hosts/domains, as
described in articles such as:
https://www.uber.com/en-GB/blog/denial-by-dns/
https://medium.com/@amirilovic/how-to-fix-node-dns-issues-5d4ec2e12e95
https://medium.com/swlh/solving-node-dns-issues-and-other-things-5051d8526cac

I'm trying the 'got' npm module for these downloads because,
unlike the request module currently used, it offers a DNS cache.
I'm hoping that caching and reusing the DNS lookups will avoid
the frequency of the ETIMEDOUT and ESOCKETTIMEDOUT errors.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>