fix: protect against bad training data
A user had put an empty string in their training data from
their Scratch project. This commit protects against this - both
by stopping Scratch from doing this again in the future, and by
filtering out empty strings before using it in training. This is
necessary because Watson Assistant rejects training sets that
includes empty strings.

More generally, as the training data is being provided by clients
it is better to be less trusting in any of the data provided.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>