feat: local project support for numbers projects (#493)
This commit represents a re-write of the support for numbers
projects. It has two objectives:
- enable storing training data in indexeddb storage
  (This is optional, and storing training data in cloud PgSQL
   remains an option)
- move running the model to the browser
  (This is required - when this goes live, models will no longer
   be hosted in the cloud.)

Models will still be created in the cloud, but they will only be
stored for a very short period. Browsers will need to download
and store the model locally.

This will represent a major reduction in cloud workload, as
all model inferencing for non-text projects will now be moved off
of the site servers.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>