chore: migrate numbers from TFDF to YDF
Tensorflow Decision Forests are being replaced by
Yggdrasil Decision Forests.

This commit migrates the support for numbers projects in MLforKids
to use this new library.

The most obvious student benefit is that there seem to be
more packages published to pip for YDF, to avoid dependency issues
that people sometimes have today with TFDF.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>