feat: allow users to choose webcam
Extending the previous contribution by adding a "change camera"
button to the webcam dialog.

This isn't supported by the webcam-directive, so I've needed to
fork the directive to add the custom video options override.
(I haven't bothered offering the modification to the original
author, as they seem to have abandoned the repo, as it has many
ignored issues and pull requests and no changes in over 7 years)

Closes: #475

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>