fix: memory leaks in tfjs image training
There were several tensors that weren't getting disposed. This
meant the GPU weight of the models page would continue to grow if
a student repeatedly re-trains a model.

I need to make equivalent fixes to the Scratch module, so that
still has a similar leak, but I don't think it's worth waiting for
that to get this fix out.

I've also improved the error message that is displayed when the
browser runs out of GPU memory during training.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>