feat: update model training page to offer simplified option
Image projects will have an additional button for training a model
that trains a smaller, simpler model.

This button will only be displayed when the new hardware
detection service recognises constrained resources, but that isn't
connected yet.

A new top-of-page notification will be displayed when a simplified
model has been trained.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>