feat: updated model controller to train simple models
Checks for reduced resources when setting up the controller for
image projects. If constrained, it sets a flag in the scope to
display the new "simplified model" button.

Also adds some additional error handling so that errors can be
displayed instead of the current "Unknown error" that is often
shown.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>