feat: improve image classifier support for low-powered devices (#523)
In the last couple months, I've seen an increased number of Sentry error 
reports during training of image classifiers, mostly from what look like 
cheap Android tablets and old Windows computers. There are a variety of 
error messages, but they all seem to suggest the problem is a lack of 
resources (e.g. GPU memory exhausted), often from students who have 
collected a very large number of training images.

This branch is an attempt to offer those users an alternate model architecture
- removing the hidden layer
- different, lighter-weight optimizer during training

It also simplifies the training:
- cpu backend, instead of webgl
- smaller batches
- smaller number of training epochs

All of this will make the model less effective and less accurate, but as a 
last resort, it's better to have a model that works at all. The new model 
training approach is optional - the user will have to click a different
"simplified ML model" button to use this approach. To avoid confusing the 
majority of users who don't need this with an extra choice to make, this 
different "simplified model" button will only be displayed when the user's 
device is observed to be low-powered (e.g. low memory)

This does not alter how the Scratch block will train image models†, so for 
now users on low powered devices will need to train their model from the 
training tool before opening Scratch. Such a device is probably better doing 
tasks (training a model and running the Scratch app) separately anyway, so 
this feels like a reasonable place to start.


† - The exception to this is the number of epochs used, as the previous 
approach was a bit backwards anyway, so taking this opportunity to fix it in 
both places.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>