fix: improve small language model support on browsers without f16
shader-f16 is a GPU feature that is required by many of the
language models that I chose to offer. While widely supported on
MacOS, a Windows user reported that their browser did not support
this. This left them with only three models to choose from, but
worse than that, trial and error was the only way for them to
realistically know which of the eight models they could use.

In this commit, I've added an f32 variant for all models that
require shader-f16 - and automatically swap those in if I detect
that shader-f16 support is missing.

This means all users should be able to choose any model, and an
appropriate option will be used.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>